+++ src/SFA.DAS.Payments.ScheduledJobs/ApprenticeshipsReferenceDataComparison/ApprenticeshipsReferenceDataComparisonTrigger.cs
@@ -4,18 +4,18 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 using SFA.DAS.Payments.ScheduledJobs.Infrastructure.IoC;
 
-namespace SFA.DAS.Payments.ScheduledJobs.ApprovalsReferenceDataComparison
+namespace SFA.DAS.Payments.ScheduledJobs.ApprenticeshipsReferenceDataComparison
 {
     [DependencyInjectionConfig(typeof(DependencyInjectionConfig))]
-    public static class ApprovalsReferenceDataComparisonTrigger
+    public static class ApprenticeshipsReferenceDataComparisonTrigger
     {
         [FunctionName("ApprovalsReferenceDataComparison")]
-        public static void Run([TimerTrigger("%ApprovalsReferenceDataComparisonSchedule%", RunOnStartup = true)]TimerInfo myTimer, [Inject]IApprovalsReferenceDataComparisonService service, ILogger log)
+        public static void Run([TimerTrigger("%ApprovalsReferenceDataComparisonSchedule%", RunOnStartup = true)]TimerInfo myTimer, [Inject]IApprenticeshipsReferenceDataComparisonService service, ILogger log)
         {
             RunApprovalsReferenceDataComparison(service, log);
         }
 
-        private static void RunApprovalsReferenceDataComparison(IApprovalsReferenceDataComparisonService service, ILogger log)
+        private static void RunApprovalsReferenceDataComparison(IApprenticeshipsReferenceDataComparisonService service, ILogger log)
         {
             try
             {
