+++ src/SFA.DAS.Payments.ScheduledJobs/ApprenticeshipsReferenceDataComparison/IApprenticeshipsReferenceDataComparisonService.cs
@@ -1,0 +1,7 @@
+﻿namespace SFA.DAS.Payments.ScheduledJobs.ApprenticeshipsReferenceDataComparison
+{
+    public interface IApprenticeshipsReferenceDataComparisonService
+    {
+        void ProcessComparison();
+    }
+}
