--- src/SFA.DAS.Payments.ScheduledJobs/ApprovalsReferenceDataComparison/IApprovalsReferenceDataComparisonService.cs
@@ -1,7 +1,0 @@
-﻿namespace SFA.DAS.Payments.ScheduledJobs.ApprovalsReferenceDataComparison
-{
-    public interface IApprovalsReferenceDataComparisonService
-    {
-        void ProcessComparison();
-    }
-}
