+++ src/SFA.DAS.Payments.ScheduledJobs/Infrastructure/IoC/Modules/FunctionsModule.cs
@@ -1,6 +1,6 @@
 ﻿using Autofac;
 using SFA.DAS.Payments.Core.Configuration;
-using SFA.DAS.Payments.ScheduledJobs.ApprovalsReferenceDataComparison;
+using SFA.DAS.Payments.ScheduledJobs.ApprenticeshipsReferenceDataComparison;
 using SFA.DAS.Payments.ScheduledJobs.AuditDataCleanUp;
 using SFA.DAS.Payments.ScheduledJobs.Infrastructure.Configuration;
 
@@ -12,7 +12,7 @@ namespace SFA.DAS.Payments.ScheduledJobs.Infrastructure.IoC.Modules
         {
             builder.RegisterType<FunctionsConfigurationHelper>().As<IConfigurationHelper>().SingleInstance();
             builder.RegisterType<AuditDataCleanUpService>().As<IAuditDataCleanUpService>().InstancePerLifetimeScope();
-            builder.RegisterType<ApprovalsReferenceDataComparisonService>().As<IApprovalsReferenceDataComparisonService>().SingleInstance();
+            builder.RegisterType<ApprenticeshipsReferenceDataComparisonService>().As<IApprenticeshipsReferenceDataComparisonService>().SingleInstance();
         }
     }
 }
